+++ src/main/java/com/kbalazsworks/stackjudge/domain/entities/Address.java
@@ -1,16 +1,15 @@
 package com.kbalazsworks.stackjudge.domain.entities;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record Address(
     Long id,
     Long companyId,
     String fullAddress,
-    BigDecimal markerLat,
-    BigDecimal markerLng,
-    BigDecimal manualMarkerLat,
-    BigDecimal manualMarkerLng,
+    Double markerLat,
+    Double markerLng,
+    Double manualMarkerLat,
+    Double manualMarkerLng,
     LocalDateTime createdAt,
     Long createdBy
 )
